Rename `_interfaces.ts` to `types.ts`
