+++ src/lib/buffer.ts
@@ -6,7 +6,7 @@ import { NotImplementedError } from "./_errors/not-implemented";
 import { UnknownFormatError } from "./_errors/unknown-format";
 import { WrongTypeError } from "./_errors/wrong-type";
 import { lazyProperties } from "./_helpers/lazy-properties";
-import { Lazy, SerializableType, VersionedType } from "./_interfaces";
+import { Lazy, SerializableType, VersionedType } from "./types";
 
 export type Name = "buffer";
 export const name: Name = "buffer";
