+++ src/lib/document.ts
@@ -7,7 +7,7 @@ import { NullPropertyError } from "./_errors/null-property";
 import { WrongTypeError } from "./_errors/wrong-type";
 import { lazyProperties } from "./_helpers/lazy-properties";
 import { CaseStyle, rename } from "./_helpers/rename";
-import { Lazy, SerializableType, Type as KryoType, VersionedType } from "./_interfaces";
+import { Lazy, SerializableType, Type as KryoType, VersionedType } from "./types";
 
 export type Name = "document";
 export const name: Name = "document";
