+++ src/lib/float64.ts
@@ -2,7 +2,7 @@ import { Incident } from "incident";
 import { UnknownFormatError } from "./_errors/unknown-format";
 import { WrongTypeError } from "./_errors/wrong-type";
 import { lazyProperties } from "./_helpers/lazy-properties";
-import { Lazy, SerializableType, VersionedType } from "./_interfaces";
+import { Lazy, SerializableType, VersionedType } from "./types";
 
 export type Name = "float64";
 export const name: Name = "float64";
