+++ src/lib/integer.ts
@@ -3,7 +3,7 @@ import { InvalidIntegerError } from "./_errors/invalid-integer";
 import { UnknownFormatError } from "./_errors/unknown-format";
 import { WrongTypeError } from "./_errors/wrong-type";
 import { lazyProperties } from "./_helpers/lazy-properties";
-import { Lazy, SerializableType, VersionedType } from "./_interfaces";
+import { Lazy, SerializableType, VersionedType } from "./types";
 
 export type Name = "int";
 export const name: Name = "int";
