+++ src/lib/map.ts
@@ -3,7 +3,7 @@ import { NotImplementedError } from "./_errors/not-implemented";
 import { UnknownFormatError } from "./_errors/unknown-format";
 import { WrongTypeError } from "./_errors/wrong-type";
 import { lazyProperties } from "./_helpers/lazy-properties";
-import { Lazy, SerializableType, VersionedType } from "./_interfaces";
+import { Lazy, SerializableType, VersionedType } from "./types";
 
 export type Name = "map";
 export const name: Name = "map";
