+++ src/lib/null.ts
@@ -1,6 +1,6 @@
 import { UnknownFormatError } from "./_errors/unknown-format";
 import { WrongTypeError } from "./_errors/wrong-type";
-import { SerializableType, VersionedType } from "./_interfaces";
+import { SerializableType, VersionedType } from "./types";
 
 export type Name = "null";
 export const name: Name = "null";
