+++ src/lib/ucs2-string.ts
@@ -7,7 +7,7 @@ import { PatternNotMatchedError } from "./_errors/pattern-not-matched";
 import { UnknownFormatError } from "./_errors/unknown-format";
 import { WrongTypeError } from "./_errors/wrong-type";
 import { lazyProperties } from "./_helpers/lazy-properties";
-import { Lazy, SerializableType, VersionedType } from "./_interfaces";
+import { Lazy, SerializableType, VersionedType } from "./types";
 
 export type Name = "ucs2-string";
 export const name: Name = "ucs2-string";
